requirements.txt
paho-mqtt
python-dotenv
aiomysql
sqlalchemy
httpx
pygelf
fastapi
uvicorn[standard]
python-multipart
aiosqlite
pywin32